fix(typescript): Fix optional chaining unused expr
The default ESLint `no-unused-expressions` rule cannot handle optional
chaining in TypeScript, for example:

```
onChangeHook?.(nextValue);
```

In typescript-eslint/typescript-eslint#1175 the base rule was duplicated
in to a TypeScript-specific one that handles optional chaining
correctly. Presumably this will eventually be merged back to the base
rule as this is now part of the ECMAScript standard.