fix(deps): Update ESLint, plugins & TypeScript
This is mostly to get access to the improvements in
`typescript/eslint@3`:
https://github.com/typescript-eslint/typescript-eslint/releases/tag/v3.0.0

This leaves `babel-eslint` alone as it now requires a babel config in
`babel-eslint@11`.

As a side-effect this also gets rid of a lot of peer dep spam for
plugins that didn't include ESLint >= 7.

BREAKING CHANGE: Drops support for Node 8
BREAKING CHANGE: Drops support for TypeScript 3.2