fix(deps): Update ESLint plugins
The last update in #50 was probably a bit rushed and caught
`typescript-eslint@3.0.0`. This release had a lot of really good
changes, but also some rough edges due to the new rules lacking real
world testing.

In particular, typescript-eslint/typescrpit-eslint#2084 made any
exported function require all of its inner functions to be fully typed.
This is particularly difficult with GraphQL servers as it requires every
single resolver function to be fully typed. We effectively halted the
rollout of `eslint-config-seek@7.0.0.` waiting for this to be fixed.
This has now been included in `typescript-eslint@3.1.0` along with
some other hotfixes.

It looks like we also hit some near-.0 releases for the other plugins so
update those as well even though they've caused no known problems.