Migrate to Changesets
It's looking like our config will require numerous breaking changes to
get to ESLint v8 compatibility (#62). If we stuck with semantic-release
we'd either need to get creative with branching, have one mega PR or
release a bunch of intermediary major versions that add little value.

Instead, switch to Changesets. This is used in a number of our OSS repos
and makes it easier to bundle multiple changes into a release. We
support previewing the release notes with and without a `GITHUB_TOKEN`
via `yarn version`.